+++ __tests__/Button.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, fireEvent } from "@testing-library/react-native";
-import Button from "../src/components/Button";
+import Button from "../src/components/buttons/Button";
 
 describe("Button", () => {
   it("renders correctly with a title", () => {
