+++ __tests__/OrderCard.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, fireEvent } from "@testing-library/react-native";
-import OrderCard from "../src/components/OrderCard";
+import OrderCard from "../src/components/cards/OrderCard";
 import { Order, OrderStatus } from "../src/types/Order";
 import { Item } from "../src/types/Item";
