+++ __tests__/OrderMenu.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { screen, render, fireEvent } from "@testing-library/react-native";
-import OrderMenu from "../src/app/OrderMenu";
+import OrderMenu from "../src/app/order/OrderMenu";
 import { productButtons } from "../src/types/ProductButtons";
 import { View, Text } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
