+++ src/components/buttons/Button.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { TouchableOpacity, Text, StyleSheet } from "react-native";
-import { buttonStyles } from "../styles/ButtonStyles";
+import { buttonStyles } from "../../styles/ButtonStyles";
 
 interface ButtonProps {
   title: string;
